requirements.txt
# For collectivo
Django>=4.1.1,<4.2
djangorestframework>=3.13.1,<3.14
django-filter==22.1
drf-spectacular>=0.24.2,<0.25
python-keycloak>=2.5.0,<2.6
PyJWT>=2.5.0
html2text>=2019.8.11<2019.9
celery>=5.2.7,<5.3
redis>=4.4.1,<4.5

# For docker
psycopg2>=2.9.3,<2.10

# For development
django-cors-headers==3.12.0
flake8>=5.0.4,<5.1
flake8-bandit>=4.1.1
flake8-docstrings>=1.6.0

# For deployment
gunicorn>=20.1.0,<20.2
